fixed decimal operations

+++ d03/calculator_project/my_arithmetic.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python3
+from decimal import *
 #----------------------------------------------------------------------
 def add(nbrs):
     result = 0
     for nb in nbrs:
         if "." in nb:
-            result += float(nb)
+            result += Decimal(nb)
         else:
             result += int(nb) 
     return result
@@ -19,13 +20,13 @@ def division(nbrs):
         for nb in nbrs:
             if first == 0:
                 if "." in nb:
-                    result += float(nb)
+                    result += Decimal(nb)
                 else :
                     result += int(nb)
                 first = 1
             else:
                 if "." in nb:
-                    result /= float(nb)
+                    result /= Decimal(nb)
                 else:
                     result /= int(nb)
     return result
@@ -37,7 +38,7 @@ def multiply(nbrs):
     result = 1
     for nb in nbrs:
         if "." in nb:
-            result *= float(nb)
+            result *= Decimal(nb)
         else:
             result *= int(nb)
     return result
@@ -49,13 +50,13 @@ def subtract(nbrs):
     for nb in nbrs:
         if first == 0:
             if "." in nb:
-                result += float(nb)
+                result += Decimal(nb)
             else:
                 result += int(nb)
             first = 1
         else:
             if "." in nb:
-                result -= float(nb)
+                result -= Decimal(nb)
             else:
                 result -= int(nb)
     return result
